refactor: change directories naming
