+++ src/components/ActivityCreation/fields/ImageDropzone.tsx
