+++ src/components/ActivityCreation/sections/ActivityContentFormSection.tsx
@@ -18,11 +18,11 @@ import { H2, H4, P } from '@components/ui/typography';
 import { createActivityFormSchema } from '@lib/definitions';
 import { UseFormReturn } from 'react-hook-form';
 import { z } from 'zod';
-import ActivityTimePicker from '../ui/AcitivityTimePicker';
-import ActivityCreationMap from '../ui/ActivityCreationMap';
-import ActivityDatePicker from '../ui/ActivityDatePicker';
-import ImageDropzone from '../ui/ImageDropzone';
-import RichTextEditor from '../ui/RichTextEditor';
+import ActivityTimePicker from '../fields/AcitivityTimePicker';
+import ActivityCreationMap from '../fields/ActivityCreationMap';
+import ActivityDatePicker from '../fields/ActivityDatePicker';
+import ImageDropzone from '../fields/ImageDropzone';
+import RichTextEditor from '../fields/RichTextEditor';
 
 type FormSchema = z.infer<typeof createActivityFormSchema>;
 
@@ -179,6 +179,53 @@ const ActivityContentFormSection = ({
             </FormItem>
           )}
         />
+        <ActivityCreationMap
+          setLat={(lat: number) => {
+            setValue('lat', lat);
+          }}
+          setLng={(lng: number) => {
+            setValue('lng', lng);
+          }}
+          setAddress={(address: string) => {
+            setValue('address', address);
+          }}
+        />
+        <FormField
+          control={form.control}
+          name="lng"
+          render={({ field }) => (
+            <FormItem className="grow">
+              <FormControl>
+                <input type="hidden" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <FormField
+          control={form.control}
+          name="lat"
+          render={({ field }) => (
+            <FormItem className="grow">
+              <FormControl>
+                <input type="hidden" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <FormField
+          control={form.control}
+          name="address"
+          render={({ field }) => (
+            <FormItem className="grow">
+              <FormControl>
+                <input type="hidden" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
       </div>
       <Separator className="h-[0.5px]" />
       <div className="space-y-6">
@@ -311,53 +358,6 @@ const ActivityContentFormSection = ({
               <FormDescription className="text-primary-light">
                 啟用週期設定將使活動固定舉辦
               </FormDescription>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <ActivityCreationMap
-          setLat={(lat: number) => {
-            setValue('lat', lat);
-          }}
-          setLng={(lng: number) => {
-            setValue('lng', lng);
-          }}
-          setAddress={(address: string) => {
-            setValue('address', address);
-          }}
-        />
-        <FormField
-          control={form.control}
-          name="lng"
-          render={({ field }) => (
-            <FormItem className="grow">
-              <FormControl>
-                <input type="hidden" {...field} />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <FormField
-          control={form.control}
-          name="lat"
-          render={({ field }) => (
-            <FormItem className="grow">
-              <FormControl>
-                <input type="hidden" {...field} />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <FormField
-          control={form.control}
-          name="address"
-          render={({ field }) => (
-            <FormItem className="grow">
-              <FormControl>
-                <input type="hidden" {...field} />
-              </FormControl>
               <FormMessage />
             </FormItem>
           )}
